# api.yaml
openapi: 3.0.3
info:
  title: Capacity Planning API
  version: 1.0.0
  description: API визуального планировщика ресурсов для студенческих проектных команд
servers:
  - url: http://localhost:8000
...
paths:
  /students/profile:
    get:
      tags: [Students]
      summary: Получить профиль текущего студента
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль текущего студента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /students:
    get:
      tags: [Students]
      summary: Получить список студентов
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/StudentEmailFilter'
        - $ref: '#/components/parameters/StudentFirstNameFilter'
        - $ref: '#/components/parameters/StudentLastNameFilter'
        - $ref: '#/components/parameters/StudentSkillsFilter'
      responses:
        '200':
          description: Список студентов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /students/{student_id}:
    get:
      tags: [Students]
      summary: Получить информацию о студенте
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Информация о студенте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Students]
      summary: Обновить данные студента
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        $ref: '#/components/requestBodies/StudentUpdateBody'
      responses:
        '200':
          description: Данные студента обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Students]
      summary: Удалить студента
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '204':
          description: Студент удален
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects:
    get:
      tags: [Projects]
      summary: Получить список проектов
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/ProjectNameFilter'
        - $ref: '#/components/parameters/ProjectOwnerStudentIdFilter'
        - $ref: '#/components/parameters/ProjectCreatedFromFilter'
        - $ref: '#/components/parameters/ProjectCreatedToFilter'
      responses:
        '200':
          description: Список проектов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Projects]
      summary: Создать проект
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/ProjectCreateBody'
      responses:
        '201':
          description: Проект создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/join:
    post:
      tags: [Projects]
      summary: Вступить в проект
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/ProjectJoinBody'
      responses:
        '201':
          description: Студент добавлен в проект
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /projects/{project_id}:
    get:
      tags: [Projects]
      summary: Получить информацию о проекте
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Информация о проекте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Projects]
      summary: Обновить проект
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        $ref: '#/components/requestBodies/ProjectUpdateBody'
      responses:
        '200':
          description: Проект обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Projects]
      summary: Удалить проект
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Проект удален
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/members:
    get:
      tags: [ProjectMembers]
      summary: Получить список участников проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/ProjectMemberStudentIdFilter'
        - $ref: '#/components/parameters/ProjectMemberRoleFilter'
        - $ref: '#/components/parameters/ProjectMemberIsActiveFilter'
        - $ref: '#/components/parameters/ProjectMemberJoinDateFromFilter'
        - $ref: '#/components/parameters/ProjectMemberJoinDateToFilter'
      responses:
        '200':
          description: Список участников проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [ProjectMembers]
      summary: Добавить участника в проект
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        $ref: '#/components/requestBodies/ProjectMemberCreateByProjectBody'
      responses:
        '201':
          description: Участник добавлен в проект
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{project_id}/members/{student_id}:
    get:
      tags: [ProjectMembers]
      summary: Получить информацию об участнике проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Информация об участнике проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [ProjectMembers]
      summary: Обновить данные участника проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        $ref: '#/components/requestBodies/ProjectMemberUpdateByProjectBody'
      responses:
        '200':
          description: Данные участника проекта обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [ProjectMembers]
      summary: Удалить участника из проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/StudentId'
      responses:
        '204':
          description: Участник удален из проекта
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/teams:
    get:
      tags: [Teams]
      summary: Получить список команд проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Список команд проекта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Teams]
      summary: Создать команду внутри проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        $ref: '#/components/requestBodies/TeamCreateByProjectBody'
      responses:
        '201':
          description: Команда создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /projects/{project_id}/teams/{team_id}:
    get:
      tags: [Teams]
      summary: Получить информацию о команде проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Информация о команде
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Teams]
      summary: Обновить команду проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        $ref: '#/components/requestBodies/TeamUpdateBody'
      responses:
        '200':
          description: Команда обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Teams]
      summary: Удалить команду из проекта
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Команда удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/teams/{team_id}/members:
    get:
      tags: [TeamMembership]
      summary: Получить список участников конкретной команды
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Список участников команды
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [TeamMembership]
      summary: Добавить участника проекта в команду
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        $ref: '#/components/requestBodies/TeamMembershipCreateByTeamBody'
      responses:
        '201':
          description: Участник добавлен в команду
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Номер страницы
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: Количество элементов на странице
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    StudentEmailFilter:
      name: email
      in: query
      required: false
      description: Фильтр по email
      schema:
        type: string
        format: email
    StudentFirstNameFilter:
      name: first_name
      in: query
      required: false
      description: Фильтр по имени
      schema:
        type: string
    StudentLastNameFilter:
      name: last_name
      in: query
      required: false
      description: Фильтр по фамилии
      schema:
        type: string
    StudentSkillsFilter:
      name: skills
      in: query
      required: false
      description: Фильтр по навыкам
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    StudentResponse:
      type: object
      required:
        - id
        - email
        - first_name
        - last_name
        - registered_at
      properties:
        id:
          type: integer
          description: Уникальный идентификатор студента
          example: 1
        email:
          type: string
          format: email
          description: Email студента
          example: student@example.com
        first_name:
          type: string
          description: Имя студента
          example: Иван
        last_name:
          type: string
          description: Фамилия студента
          example: Иванов
        registered_at:
          type: string
          format: date-time
          description: Дата регистрации
          example: '2026-03-13T10:00:00Z'
        skills:
          type: string
          nullable: true
          description: Навыки студента
          example: Python, FastAPI, PostgreSQL
...
    StudentListResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
...